# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuBox BE",
    "version": "1.0.0",
    "description": "REST API for the live Q&A question pool platform. Participants join sessions, submit and upvote questions; moderators curate the pool. Real-time updates are delivered over WebSocket (see WebSocket Events section)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/sessions": {
      "post": {
        "tags": ["Sessions"],
        "summary": "Create a new session",
        "description": "Creates a new Q&A session. The session starts in `created` status and must be explicitly started before it accepts questions.",
        "operationId": "createSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" },
              "example": {
                "title": "AMA: Product Roadmap Q3",
                "description": "Ask us anything about the upcoming roadmap.",
                "anonymousAllowed": true,
                "hostId": "host-uuid-001"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          }
        }
      }
    },
    "/sessions/{id}/start": {
      "post": {
        "tags": ["Sessions"],
        "summary": "Start a session",
        "description": "Transitions the session from `created` to `open`. Once open, participants can submit and upvote questions.",
        "operationId": "startSession",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HostActionRequest" },
              "example": { "hostId": "host-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session started. Status is now `open`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/TransitionError" }
        }
      }
    },
    "/sessions/{id}/end": {
      "post": {
        "tags": ["Sessions"],
        "summary": "End a session",
        "description": "Transitions the session from `open` to `closed`. No new questions or upvotes are accepted after this point. A `session_closed` WebSocket event is broadcast to all connected clients.",
        "operationId": "endSession",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HostActionRequest" },
              "example": { "hostId": "host-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session ended. Status is now `closed`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/TransitionError" }
        }
      }
    },
    "/sessions/{joinCode}": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Join a session by join code",
        "description": "Looks up a session by its join code and returns the session details, a participant ID, and the current list of visible questions. If `anonymousAllowed` is true and no `participantId` is provided, a new anonymous UUID is assigned. Re-joining with the same join code returns the same participant ID.",
        "operationId": "joinSession",
        "parameters": [
          { "$ref": "#/components/parameters/JoinCode" }
        ],
        "responses": {
          "200": {
            "description": "Successfully joined the session.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JoinResult" }
              }
            }
          },
          "404": {
            "description": "Invalid or expired join code.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Session not found" }
              }
            }
          }
        }
      }
    },
    "/sessions/{id}/questions": {
      "post": {
        "tags": ["Questions"],
        "summary": "Submit a question",
        "description": "Submits a new question to an open session. The question is assigned a UUID and timestamp, and its initial status is `pending`. Questions submitted to a closed session are rejected.",
        "operationId": "submitQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitQuestionRequest" },
              "example": {
                "participantId": "anon-uuid-abc123",
                "text": "What is the biggest challenge you faced this quarter?"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question submitted. Status is `pending`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "400": {
            "description": "Validation error — text is empty or exceeds 300 characters.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" },
                "example": {
                  "error": "Question text must be between 1 and 300 characters",
                  "field": "text"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Session is closed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Session is closed" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Questions"],
        "summary": "Get visible questions for a session",
        "description": "Returns all `approved` and `pinned` questions for the session, ordered by: pinned first, then approved sorted by upvote count descending, ties broken by submission timestamp ascending.",
        "operationId": "getVisibleQuestions",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "responses": {
          "200": {
            "description": "Ordered list of visible questions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Question" }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{id}/questions/all": {
      "get": {
        "tags": ["Moderation"],
        "summary": "Get all questions for a session (host/moderator view)",
        "description": "Returns all questions for the session regardless of status — including `pending`, `rejected`, `approved`, `pinned`, and `answered`. Intended for the host or moderator to review and act on pending questions. Results are ordered the same way as the visible endpoint.",
        "operationId": "getAllQuestions",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "responses": {
          "200": {
            "description": "All questions for the session in display order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Question" }
                }
              }
            }
          }
        }
      }
    },
    "/questions/{id}/approve": {
      "patch": {
        "tags": ["Moderation"],
        "summary": "Approve a question",
        "description": "Transitions a `pending` question to `approved`. If the question is not in `pending` status, a 409 reconfirmation response is returned with the current status — the caller must re-send to confirm the action.",
        "operationId": "approveQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModeratorActionRequest" },
              "example": { "moderatorId": "mod-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question approved. Status is now `approved`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/ReconfirmationRequired" }
        }
      }
    },
    "/questions/{id}/reject": {
      "patch": {
        "tags": ["Moderation"],
        "summary": "Reject a question",
        "description": "Transitions a `pending` question to `rejected`. If the question is not in `pending` status, a 409 reconfirmation response is returned with the current status.",
        "operationId": "rejectQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModeratorActionRequest" },
              "example": { "moderatorId": "mod-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question rejected. Status is now `rejected`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/ReconfirmationRequired" }
        }
      }
    },
    "/questions/{id}/pin": {
      "patch": {
        "tags": ["Moderation"],
        "summary": "Pin a question",
        "description": "Transitions a question to `pinned` status regardless of its current status. Pinned questions appear at the top of the visible question list.",
        "operationId": "pinQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModeratorActionRequest" },
              "example": { "moderatorId": "mod-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question pinned. Status is now `pinned`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/questions/{id}/unpin": {
      "patch": {
        "tags": ["Moderation"],
        "summary": "Unpin a question",
        "description": "Transitions a `pinned` question back to `approved`. The question remains visible but loses its pinned priority in the ordering. Returns 409 if the question is not currently `pinned`.",
        "operationId": "unpinQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModeratorActionRequest" },
              "example": { "moderatorId": "mod-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question unpinned. Status is now `approved`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/TransitionError" }
        }
      }
    },
    "/questions/{id}/answer": {
      "patch": {
        "tags": ["Moderation"],
        "summary": "Mark a question as answered",
        "description": "Transitions a question to `answered` status. Only allowed from `approved` or `pinned` status. Attempting from any other status returns a 409 transition error.",
        "operationId": "markAnswered",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModeratorActionRequest" },
              "example": { "moderatorId": "mod-uuid-001" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question marked as answered. Status is now `answered`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Question" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/TransitionError" }
        }
      }
    },
    "/questions/{id}/upvote": {
      "post": {
        "tags": ["Upvotes"],
        "summary": "Upvote a question",
        "description": "Records an upvote from a participant on an eligible question. Each participant may upvote a given question at most once. Only `approved` and `pinned` questions in an open session are eligible.",
        "operationId": "upvoteQuestion",
        "parameters": [
          { "$ref": "#/components/parameters/QuestionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpvoteRequest" },
              "example": {
                "participantId": "anon-uuid-abc123",
                "sessionId": "session-uuid-001"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upvote recorded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Upvote" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Participant has already upvoted this question.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Already upvoted" }
              }
            }
          },
          "422": {
            "description": "Question is not eligible for upvoting (status is `pending`, `rejected`, or `answered`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Question is not eligible for upvoting" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "SessionId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Session UUID",
        "schema": { "type": "string", "format": "uuid" }
      },
      "JoinCode": {
        "name": "joinCode",
        "in": "path",
        "required": true,
        "description": "Short alphanumeric join code assigned to the session on creation (e.g. `ABC123`)",
        "schema": { "type": "string", "example": "ABC123" }
      },
      "QuestionId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Question UUID",
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "schemas": {
      "Session": {
        "type": "object",
        "required": ["id", "joinCode", "title", "hostId", "status", "anonymousAllowed", "createdAt"],
        "properties": {
          "id":               { "type": "string", "format": "uuid" },
          "joinCode":         { "type": "string", "example": "ABC123", "description": "Short unique code participants use to join" },
          "title":            { "type": "string", "example": "AMA: Product Roadmap Q3" },
          "description":      { "type": "string", "nullable": true },
          "hostId":           { "type": "string", "format": "uuid" },
          "status":           { "type": "string", "enum": ["created", "open", "closed"] },
          "anonymousAllowed": { "type": "boolean" },
          "createdAt":        { "type": "string", "format": "date-time" },
          "startedAt":        { "type": "string", "format": "date-time", "nullable": true },
          "endedAt":          { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "Question": {
        "type": "object",
        "required": ["id", "sessionId", "participantId", "text", "status", "upvoteCount", "submittedAt", "lastModifiedAt"],
        "properties": {
          "id":             { "type": "string", "format": "uuid" },
          "sessionId":      { "type": "string", "format": "uuid" },
          "participantId":  { "type": "string", "format": "uuid" },
          "text":           { "type": "string", "minLength": 1, "maxLength": 300 },
          "status":         { "type": "string", "enum": ["pending", "approved", "rejected", "answered", "pinned"] },
          "upvoteCount":    { "type": "integer", "minimum": 0 },
          "submittedAt":    { "type": "string", "format": "date-time" },
          "lastModifiedAt": { "type": "string", "format": "date-time" },
          "lastModifiedBy": { "type": "string", "format": "uuid", "nullable": true, "description": "Moderator ID of the last status change" }
        }
      },
      "Upvote": {
        "type": "object",
        "required": ["questionId", "participantId", "sessionId", "createdAt"],
        "properties": {
          "questionId":    { "type": "string", "format": "uuid" },
          "participantId": { "type": "string", "format": "uuid" },
          "sessionId":     { "type": "string", "format": "uuid" },
          "createdAt":     { "type": "string", "format": "date-time" }
        }
      },
      "JoinResult": {
        "type": "object",
        "required": ["session", "participantId", "questions"],
        "properties": {
          "session":       { "$ref": "#/components/schemas/Session" },
          "participantId": {
            "type": "string",
            "format": "uuid",
            "description": "Existing participant ID (on re-join) or newly assigned anonymous UUID. Store this client-side for subsequent requests."
          },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" },
            "description": "Current visible questions (approved + pinned) in display order"
          }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "required": ["title", "anonymousAllowed", "hostId"],
        "properties": {
          "title":            { "type": "string", "example": "AMA: Product Roadmap Q3" },
          "description":      { "type": "string", "nullable": true },
          "anonymousAllowed": { "type": "boolean", "default": true },
          "hostId":           { "type": "string", "format": "uuid" }
        }
      },
      "HostActionRequest": {
        "type": "object",
        "required": ["hostId"],
        "properties": {
          "hostId": { "type": "string", "format": "uuid" }
        }
      },
      "SubmitQuestionRequest": {
        "type": "object",
        "required": ["participantId", "text"],
        "properties": {
          "participantId": { "type": "string", "format": "uuid", "description": "The participant ID returned from the join endpoint" },
          "text":          { "type": "string", "minLength": 1, "maxLength": 300 }
        }
      },
      "ModeratorActionRequest": {
        "type": "object",
        "required": ["moderatorId"],
        "properties": {
          "moderatorId": { "type": "string", "format": "uuid" }
        }
      },
      "UpvoteRequest": {
        "type": "object",
        "required": ["participantId", "sessionId"],
        "properties": {
          "participantId": { "type": "string", "format": "uuid" },
          "sessionId":     { "type": "string", "format": "uuid" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "field": { "type": "string", "description": "The request field that failed validation", "example": "text" }
        }
      },
      "ReconfirmationRequiredResponse": {
        "type": "object",
        "required": ["error", "currentStatus"],
        "properties": {
          "error":         { "type": "string", "example": "Question is not in pending status (current: approved)" },
          "currentStatus": { "type": "string", "enum": ["pending", "approved", "rejected", "answered", "pinned"], "description": "The question's actual current status" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Not found" }
          }
        }
      },
      "TransitionError": {
        "description": "Invalid state transition.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Cannot transition from current status" }
          }
        }
      },
      "ReconfirmationRequired": {
        "description": "Question is not in `pending` status. The response includes the current status so the caller can decide whether to confirm the action.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ReconfirmationRequiredResponse" }
          }
        }
      }
    }
  },
...
}